ml-api/pyproject.toml
[tool.poetry]
name = "ml-api"
version = "0.1.0"
description = "API For the ML related features in SCOPE. See README.md for more information."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
# packages = [{ include = "ml_api", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = ">=0.109.2"
uvicorn = ">=0.27.1"
python-dotenv = ">=1.0.1"
pydantic-settings = ">=2.5.2"
llama-index-core = "*"
llama-index-program-openai = "*"
llama-index-llms-openai-like = "*"
llama-index-embeddings-text-embeddings-inference = "*"
llama-index-vector-stores-qdrant = "*"
llama-index-readers-file = "*"
pydantic = "^2.10.6"
docx2txt = "^0.8"
llama-index-multi-modal-llms-openai = "^0.5.0"
weave = "^0.51.39"
mlflow = "^2.21.2"
pdf2image = "^1.17.0"
poppler-utils = "^0.1.0"
aspose-words = "^25.4.0"


[tool.poetry.group.dev.dependencies]
black = ">=24.2.0"

[tool.poetry.group.scraper.dependencies]
dask = "^2024.9.0"
beautifulsoup4 = "^4.12.3"

[tool.poetry.group.test.dependencies]
pytest = ">=8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
